+++ tests/helpers/test_server.rs
@@ -5,7 +5,6 @@ use async_std::io::prelude::ReadExt;
 use diesel::PgConnection;
 use http_service::Response;
 use http_service_mock::{make_server, TestBackend};
-use realworld_tide::configuration::Settings;
 use serde_json::Value;
 use tide::server::Service;
 
@@ -14,11 +13,6 @@ pub type TestServer = TestBackend<Service<Repo<PgConnection>>>;
 pub fn new(repo: Repo<PgConnection>) -> TestServer {
     let app = get_app(repo);
     make_server(app.into_http_service()).unwrap()
-}
-
-pub fn get_repo() -> Repo<PgConnection> {
-    let settings = Settings::new().expect("Failed to load configuration");
-    return Repo::new(&settings.database.connection_string());
 }
 
 pub async fn response_json(mut res: Response) -> Value {
